Us 437011 integrating scan central (#67)
* Added Scan Central Packing Code

* General layout done prepared for finalizing ScanCentral

* Added build commands for scan central

* Made changes to remove duplicate stuff

* Changed error message as per comments

* Added some logging

* Restructured some piece of Scan Central code

* Uncommented deltion part

+++ src/main/java/org/jenkinsci/plugins/fodupload/SharedUploadBuildStep.java
@@ -1,13 +1,17 @@
 package org.jenkinsci.plugins.fodupload;
 
-import java.io.File;
-import java.io.IOException;
-import java.io.PrintStream;
+import java.io.*;
+import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.util.ArrayList;
 import java.util.EnumSet;
 import java.util.List;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 import com.cloudbees.plugins.credentials.CredentialsProvider;
 
+import org.apache.maven.artifact.versioning.ComparableVersion;
 import org.jenkinsci.plugins.fodupload.controllers.ApplicationsController;
 import org.jenkinsci.plugins.fodupload.controllers.StaticScanController;
 import org.jenkinsci.plugins.fodupload.models.AuthenticationModel;
@@ -474,33 +478,50 @@ public class SharedUploadBuildStep {
 
                     technologyStack = staticScanSetup.getTechnologyStack();
                 }
-                String scanCentralPath = GlobalConfiguration.all().get(FodGlobalDescriptor.class).getScanCentralPath();
 
                 FilePath workspaceModified = new FilePath(workspace, model.getSrcLocation());
-                // zips the file in a temporary location
-                File payload = Utils.createZipFile(technologyStack, workspaceModified, logger);
-                if (payload.length() == 0) {
-
-                    boolean deleteSuccess = payload.delete();
-                    if (!deleteSuccess) {
-                        logger.println("Unable to delete empty payload.");
+                File payload;
+                if (model.getSelectedScanCentralBuildType().equalsIgnoreCase(FodEnums.SelectedScanCentralBuildType.None.toString())) {
+                    // zips the file in a temporary location
+                    payload = Utils.createZipFile(technologyStack, workspaceModified, logger);
+                    if (payload.length() == 0) {
+                        boolean deleteSuccess = payload.delete();
+                        if (!deleteSuccess) {
+                            logger.println("Unable to delete empty payload.");
+                        }
+                        logger.println("Source is empty for given Technology Stack and Language Level.");
+                        build.setResult(Result.FAILURE);
+                        return;
                     }
+                } else {
+                    FilePath scanCentralPath = new FilePath(new File(GlobalConfiguration.all().get(FodGlobalDescriptor.class).getScanCentralPath()));
+                    logger.println("Scan Central Path : " + scanCentralPath);
+                    Path scPackPath = packageScanCentral(workspaceModified, scanCentralPath, workspace, model, logger, build);
+                    logger.println("Packaged File Output Path : " + scPackPath);
 
-                    logger.println("Source is empty for given Technology Stack and Language Level.");
-                    build.setResult(Result.FAILURE);
-                    return;
+                    if (scPackPath != null) {
+                        payload = new File(scPackPath.toString());
+
+                        if (!payload.exists()) {
+                            build.setResult(Result.FAILURE);
+                            return;
+                        }
+                    } else {
+                        logger.println("Scan Central package output not found.");
+                        build.setResult(Result.FAILURE);
+                        return;
+                    }
                 }
 
                 model.setPayload(payload);
-
                 String notes = String.format("[%d] %s - Assessment submitted from Jenkins FoD Plugin",
                         build.getNumber(),
                         build.getDisplayName());
 
                 StartScanResponse scanResponse = staticScanController.startStaticScan(releaseId, model, notes);
                 boolean deleted = payload.delete();
-
                 boolean isWarningSettingEnabled = model.getInProgressBuildResultType().equalsIgnoreCase(InProgressBuildResultType.WarnBuild.getValue());
+
                 /**
                  * If(able to contact api) {
                  *      if(Scan is allowed to start && the uploaded file is deleted) {
@@ -586,4 +607,192 @@ public class SharedUploadBuildStep {
     public int setScanId(int newScanId) {
         return scanId = newScanId;
     }
+
+    private Path packageScanCentral(FilePath srcLocation, FilePath scanCentralLocation, FilePath outputLocation, JobModel job, PrintStream logger, Run<?, ?> build) {
+        BufferedReader stdInputVersion = null, stdInput = null;
+
+        try {
+            //version check
+            logger.println("Checking ScanCentralVersion");
+            String scanCentralbatLocation = Paths.get(String.valueOf(scanCentralLocation)).resolve("scancentral.bat").toString();
+            ArrayList scanCentralVersionCommandList = new ArrayList<>();
+            scanCentralVersionCommandList.add(scanCentralbatLocation);
+            scanCentralVersionCommandList.add("--version");
+            Process pVersion = runProcessBuilder(scanCentralVersionCommandList, scanCentralLocation);
+            stdInputVersion = new BufferedReader(new InputStreamReader(
+                    pVersion.getInputStream()));
+            String versionLine = null;
+            String scanCentralVersion = null;
+            Boolean isValidVersion = false;
+
+            while ((versionLine = stdInputVersion.readLine()) != null) {
+                logger.println(versionLine);
+                if (versionLine.contains("version")) {
+
+                    Pattern versionPattern = Pattern.compile("(?<=version:  ).*");
+                    Matcher m = versionPattern.matcher(versionLine);
+
+                    if (m.find()) {
+                        scanCentralVersion = m.group().trim();
+
+                        ComparableVersion minScanCentralVersion = new ComparableVersion("20.2.0.0019");
+                        ComparableVersion userScanCentralVersion = new ComparableVersion(scanCentralVersion);
+
+                        if (userScanCentralVersion.compareTo(minScanCentralVersion) < 0) {
+                            logger.println("The supplied scan central version is outdated . Please upgrade to higher version and try again !!");
+                            build.setResult(Result.FAILURE);
+                            return null;
+                        }
+                        logger.println("ScanCentral Version " + scanCentralVersion);
+                        isValidVersion = true;
+                        break;
+                    }
+                }
+            }
+            if (versionLine.contains("version")) {
+                Path outputZipFolderPath = Paths.get(String.valueOf(outputLocation)).resolve("output.zip");
+                ArrayList scanCentralPackageCommandList = new ArrayList<>();
+                scanCentralPackageCommandList.add(scanCentralbatLocation);
+                scanCentralPackageCommandList.add("package");
+                scanCentralPackageCommandList.add("--bt");
+
+                FodEnums.SelectedScanCentralBuildType buildType = FodEnums.SelectedScanCentralBuildType.valueOf(model.getSelectedScanCentralBuildType());
+
+                switch (buildType) {
+                    case Gradle:
+                        scanCentralPackageCommandList.add("gradle");
+                        if (model.getScanCentralIncludeTests()) scanCentralPackageCommandList.add("--include-test");
+                        if (model.getScanCentralSkipBuild()) scanCentralPackageCommandList.add("--skipBuild");
+                        if (!Utils.isNullOrEmpty(model.getScanCentralBuildCommand())) {
+                            scanCentralPackageCommandList.add("--build-command");
+                            scanCentralPackageCommandList.add(model.getScanCentralBuildCommand());
+                        }
+                        if (!Utils.isNullOrEmpty(model.getScanCentralBuildFile())) {
+                            scanCentralPackageCommandList.add("--build-file");
+                            scanCentralPackageCommandList.add("\"" + model.getScanCentralBuildFile() + "\"");
+                        }
+                        break;
+                    case Maven:
+                        scanCentralPackageCommandList.add("mvn");
+                        if (model.getScanCentralIncludeTests()) scanCentralPackageCommandList.add("--include-test");
+                        if (model.getScanCentralSkipBuild()) scanCentralPackageCommandList.add("--skipBuild");
+                        if (!Utils.isNullOrEmpty(model.getScanCentralBuildCommand())) {
+                            scanCentralPackageCommandList.add("--build-command");
+                            scanCentralPackageCommandList.add(model.getScanCentralBuildCommand());
+                        }
+                        if (!Utils.isNullOrEmpty(model.getScanCentralBuildFile())) {
+                            scanCentralPackageCommandList.add("--build-file");
+                            scanCentralPackageCommandList.add("\"" + model.getScanCentralBuildFile() + "\"");
+                        }
+                        break;
+                    case MSBuild:
+                        scanCentralPackageCommandList.add("msbuild");
+                        if (!Utils.isNullOrEmpty(model.getScanCentralBuildCommand())) {
+                            scanCentralPackageCommandList.add("--build-command");
+                            scanCentralPackageCommandList.add(transformMsBuildCommand(model.getScanCentralBuildCommand()));
+                        }
+                        if (!Utils.isNullOrEmpty(model.getScanCentralBuildFile())) {
+                            scanCentralPackageCommandList.add("--build-file");
+                            scanCentralPackageCommandList.add("\"" + model.getScanCentralBuildFile() + "\"");
+                        } else {
+                            logger.println("Build File is a required field for msbuild build type. Please fill in the .sln file name in the current source folder ");
+                            build.setResult(Result.FAILURE);
+                        }
+                        break;
+                    case Python:
+                        scanCentralPackageCommandList.add("none");
+                        if (!Utils.isNullOrEmpty(model.getScanCentralVirtualEnv())) {
+                            scanCentralPackageCommandList.add("--python-virtual-env");
+                            scanCentralPackageCommandList.add(model.getScanCentralVirtualEnv());
+                        }
+                        ;
+                        if (!Utils.isNullOrEmpty(model.getScanCentralRequirementFile())) {
+                            scanCentralPackageCommandList.add("--python-requirements");
+                            scanCentralPackageCommandList.add(model.getScanCentralRequirementFile());
+                        }
+                        ;
+                        if (!Utils.isNullOrEmpty(model.getScanCentralBuildToolVersion())) {
+                            scanCentralPackageCommandList.add("--python-version");
+                            scanCentralPackageCommandList.add(model.getScanCentralBuildToolVersion());
+                        }
+                        ;
+                        break;
+                    case PHP:
+                        scanCentralPackageCommandList.add("none");
+                        if (!Utils.isNullOrEmpty(model.getScanCentralBuildToolVersion())) {
+                            scanCentralPackageCommandList.add("--php-version");
+                            scanCentralPackageCommandList.add(model.getScanCentralBuildToolVersion());
+                        }
+                        ;
+                        break;
+                }
+                scanCentralPackageCommandList.add("--o");
+                scanCentralPackageCommandList.add("\"" + outputZipFolderPath.toString() + "\"");
+
+                logger.println("Packaging ScanCentral\n" + String.join(" ", scanCentralPackageCommandList));
+
+                Process scanCentralProcess = runProcessBuilder(scanCentralPackageCommandList, srcLocation);
+                stdInput = new BufferedReader(new InputStreamReader(scanCentralProcess.getInputStream()));
+                String s = null;
+                while ((s = stdInput.readLine()) != null) {
+                    logger.println(s);
+                }
+                int exitCode = scanCentralProcess.waitFor();
+                logger.println(versionLine);
+                if (exitCode != 0) {
+                    logger.println("Errors executing Scan Central. Exiting with errorcode : " + exitCode);
+                    build.setResult(Result.FAILURE);
+                } else {
+                    return outputZipFolderPath;
+                }
+            } else {
+                build.setResult(Result.FAILURE);
+                logger.println("ScanCentral not found or invalid version");
+            }
+            return null;
+        } catch (IOException | InterruptedException e) {
+            logger.println(String.format("Failed executing scan central : ", e));
+        } finally {
+            try {
+                if (stdInputVersion != null) {
+                    stdInputVersion.close();
+                }
+                if (stdInput != null) {
+                    stdInput.close();
+                }
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+        }
+        return null;
+    }
+
+    private String transformMsBuildCommand(String cmd) {
+        if (!Utils.isNullOrEmpty(cmd)) {
+            String[] arrOfCmds = cmd.split(" ");
+            StringBuilder transformedCommands = new StringBuilder();
+            for (String command : arrOfCmds) {
+                if (command.charAt(0) == '-') {
+                    command = '/' + command.substring(1);
+                }
+                transformedCommands.append(command).append(" ");
+            }
+            return transformedCommands.substring(0, transformedCommands.length() - 1);
+        }
+        return null;
+    }
+
+    private Process runProcessBuilder(ArrayList cmdList, FilePath directoryLocation) throws IOException {
+        try {
+            ProcessBuilder pb = new ProcessBuilder(cmdList);
+            pb.directory(new File(String.valueOf(directoryLocation)));
+            Process p = pb.start();
+            System.out.println(pb.redirectErrorStream());
+            pb.redirectErrorStream(true);
+            return p;
+        } catch (IOException e) {
+            throw e;
+        }
+    }
+
 }
